All implementation was moved from CompilerController to sepate class CompileManager. CodeBlockResult enum was added to simplify verification of result.
