+++ src/Axe/Views/Compiler/Index.cshtml
@@ -1,4 +1,4 @@
-@model List<Axe.Models.CodeBlock>
+@model List<Axe.ViewModels.CompilerVm.CodeBlockVm>
 @{
     ViewData["Title"] = "Compiler page";
 }
